backend/requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
docker==7.1.0
PyJWT==2.12.1
python-dotenv==1.2.1
werkzeug==3.1.8
